Separing load from utils
