+++ packages/eslint-plugin-office-addins/src/rules/no-empty-load.ts
@@ -4,8 +4,8 @@ import {
   Scope,
   Variable,
 } from "@typescript-eslint/experimental-utils/dist/ts-eslint-scope";
-import { isLoadFunction } from "../utils/utils";
 import { isGetFunction } from "../utils/getFunction";
+import { isLoadFunction } from "../utils/load";
 
 export = {
   name: "no-empty-load",
